+++ Kernel/System/Daemon/DaemonModules/SchedulerTaskWorker/Cron.pm
@@ -168,7 +168,8 @@ sub Run {
     my $Success = 1;
 
     # Check if there are errors.
-    if ( $ErrorMessage || $ConsoleCommandFailure ) {
+    # Do not log debug messages as Daemon errors. See bug#14722 (https://bugs.otrs.org/show_bug.cgi?id=14722).
+    if ( ( $ErrorMessage && $ErrorMessage !~ /Debug: /g ) || $ConsoleCommandFailure ) {
 
         $ErrorMessage //= "Console command '$Param{TaskName}' is failed.";