+++ Kernel/System/Daemon/DaemonModules/SystemConfigurationSyncManager.pm
@@ -130,7 +130,8 @@ sub Run {
     };
 
     # Check if there are errors.
-    if ( $ErrorMessage || !$Success ) {
+    # Do not log debug messages as Daemon errors. See bug#14722 (https://bugs.otrs.org/show_bug.cgi?id=14722).
+    if ( ( $ErrorMessage && $ErrorMessage !~ /Debug: /g ) || !$Success ) {
         $Self->_HandleError(
             TaskName     => 'ConfigurationDeploySync',
             TaskType     => 'SystemConfigurationSyncManager',